refactor: data set selection (#360)
Fix the code in `getStorageProviderAndValidatePayer()` to always use the
output from the previous filtering step as the input for the next
filtering step.

This does not change the worker behaviour in practice, because all data
sets in `withPaymentRail` and `withCDN` have the same payer address.
Either either all of those data sets are sanctioned or none of them are.

There is no change in tests because this commit is not changing
observable behaviour.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>